todo list, and testing info storage
Added my old to-do list txt, and i am going to actually transfer this over to the project to do lists tommorow, and then I also created my first program that writes information onto a .txt, so I am going to test and learn how to extract said information. This will be useful later on for the server-side of things.